[Add] ProductCategory 컴포넌트 생성
